HMRC-1438: Update FAQ Feedback
With the backend routes changing, the frontend routes need to change
also. They were not using standard RESTful routes, so this commit
changes them to live under `faq/feedback` and use GET and POST requests.
