+++ config/routes.rb
@@ -160,10 +160,7 @@ Rails.application.routes.draw do
     resources :results, param: :category, only: %i[create], path: 'result'
 
     get 'faq', to: 'faq#index'
-
-    get 'get_feedback', to: 'faq#get_feedback'
-
-    post 'send_feedback', to: 'faq#send_feedback', as: :send_feedback
+    resources :faq_feedback, only: %i[index create], path: 'faq/feedback'
   end
 
   match '/search', as: :perform_search, via: %i[get post], to: 'search#search'
